# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "有来管理后台 API",
        "description": "基于 ThinkPHP 8 的 RBAC 权限管理系统后端 API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "开发环境"
        }
    ],
    "paths": {
...
        "/api/v1/files": {
            "post": {
                "tags": [
                    "10.文件接口"
                ],
                "summary": "文件上传",
                "operationId": "ddff9aa00b9dca74562809b5a0c257d8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "10.文件接口"
                ],
                "summary": "文件删除",
                "operationId": "5cf6a400ffecdbf58782e7c4eead866f",
                "parameters": [
                    {
                        "name": "filePath",
                        "in": "query",
                        "description": "文件路径",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
}